Rename cst -> stage-1-cst, ast -> stage-2-ast
